# docs/openapi.yml
openapi: '3.1.0'
info:
  title: Civic Influence Graph API
  version: 0.1.0
  description: |
    RESTful JSON API for the Civic Influence Graph platform.
    All responses include provenance metadata. No identity-based targeting filters
    are exposed. Authentication required only for write operations and bulk exports.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Default API base path

# ------------------------------------------------------------------
# Common components
# ------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key passed as Bearer token
  schemas:
    PaginationMeta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        data_snapshot: { type: string, nullable: true, format: date-time }
        total_count: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
...
    GraphNode:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        name: { type: string }
        properties: { type: object }
    GraphEdge:
      type: object
      properties:
        id: { type: string, format: uuid }
        source: { type: string, format: uuid }
        target: { type: string, format: uuid }
        label: { type: string }
        properties: { type: object }
...
    SnapshotSummary:
      type: object
      properties:
        id: { type: string }
        created_at: { type: string, format: date-time }
        election_cycles: { type: array, items: { type: string } }
        data_sources: { type: array, items: { type: string } }
        record_counts: { type: object }
        format: { type: string }
        size_bytes: { type: integer }
        download_url: { type: string }
        change_log_url: { type: string }
...
paths:
...
  /graph/query:
    post:
      summary: Execute filtered graph query
      tags: [Graph]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                center_entity_id: { type: string, format: uuid }
                depth: { type: integer, default: 2 }
                filters:
                  type: object
                  properties:
                    start_date: { type: string, format: date }
                    end_date: { type: string, format: date }
                    sectors: { type: array, items: { type: string } }
                    min_amount: { type: number }
                    edge_types: { type: array, items: { type: string } }
                    jurisdictions: { type: array, items: { type: string } }
                max_nodes: { type: integer, default: 200 }
      responses:
        '200':
          description: Graph nodes and edges
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      nodes:
                        type: array
                        items: { $ref: '#/components/schemas/GraphNode' }
                      edges:
                        type: array
                        items: { $ref: '#/components/schemas/GraphEdge' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
  /graph/table:
    get:
      summary: Flat table of graph edges for export
      tags: [Graph]
      parameters:
        - name: center_entity_id
          in: query
          schema: { type: string, format: uuid }
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - name: sectors
          in: query
          schema: { type: string }
        - name: min_amount
          in: query
          schema: { type: number }
        - name: edge_types
          in: query
          schema: { type: string }
        - name: format
          in: query
          schema: { type: string, enum: [json, csv], default: json }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, maximum: 1000 }
      responses:
        '200':
          description: Table rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
...
  /export/snapshots:
    get:
      summary: List available bulk export snapshots
      tags: [Export]
      responses:
        '200':
          description: Snapshot list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      snapshots:
                        type: array
                        items: { $ref: '#/components/schemas/SnapshotSummary' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      summary: Generate a new snapshot
      tags: [Export]
      security:
        - apiKeyAuth: []
      responses:
        '200':
          description: Snapshot generated
        '401': { description: Unauthorized }
  /export/snapshots/{id}/download:
    get:
      summary: Download a bulk snapshot
      tags: [Export]
      security:
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Snapshot data
        '401': { description: Unauthorized }
        '404': { description: Not found }
  /export/snapshots/{id}/changelog:
    get:
      summary: Changelog between this and previous snapshot
      tags: [Export]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Changelog data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      changelog:
                        type: object
                        properties:
                          added: { type: integer }
                          updated: { type: integer }
                          removed: { type: integer }
                          unchanged: { type: integer }
                          entries: { type: array, items: { type: object } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '404': { description: Not found }
...
  /replication/feeds/{name}/seed:
    post:
      summary: Start or stop seeding a feed
      tags: [Replication]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action: { type: string, enum: [start, stop] }
      responses:
        '200': { description: Seed status updated }
  /replication/feeds/follow:
    post:
      summary: Follow a remote feed
      tags: [Replication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_key]
              properties:
                public_key: { type: string }
                mode: { type: string, enum: [full, sparse-jurisdiction, sparse-entity-set, snapshot-only] }
                jurisdiction: { type: string }
      responses:
        '200': { description: Following remote feed }